feat: create db models, and sync db, refactor mysql utils
