+++ server/app/models/User.model.ts
@@ -1,0 +1,15 @@
+import { DataTypes } from 'sequelize';
+import { sequelize } from '../mysql/db.create';
+
+const User = sequelize.define('User', {
+  firstName: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+  lastName: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+});
+
+export default User;
